# docs/swagger.yaml
basePath: /
definitions:
  dto.APIErrorResponse:
    properties:
      errors: {}
      message:
        example: Validation error
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.APIResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.CreateRewardRequest:
    properties:
      description:
        example: Voucher belanja dapat digunakan di minimarket rekanan
        type: string
      image:
        example: rewards/voucher50k.png
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Voucher Belanja Rp 50.000
        maxLength: 255
        minLength: 2
        type: string
      point_cost:
        example: 500
        type: integer
      stock:
        example: 20
        minimum: 0
        type: integer
    required:
    - name
    - point_cost
    - stock
    type: object
...
  dto.RewardResponse:
    properties:
      created_at:
        type: string
      description:
        example: Voucher belanja dapat digunakan di minimarket rekanan
        type: string
      id:
        example: 1
        type: integer
      image:
        example: rewards/voucher50k.png
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Voucher Belanja Rp 50.000
        type: string
      point_cost:
        example: 500
        type: integer
      stock:
        example: 20
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.UpdateRewardRequest:
    properties:
      description:
        example: Voucher belanja berlaku nasional
        type: string
      image:
        example: rewards/voucher50k_new.png
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Voucher Belanja Rp 50.000 Promo
        type: string
      point_cost:
        example: 450
        type: integer
      stock:
        example: 25
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ecopoints.test
    name: EcoPoints Dev Team
  description: REST API for EcoPoints application built with Go Gin and connected
    to existing Laravel MySQL database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EcoPoints Golang REST API
  version: "1.0"
paths:
...
  /api/v1/rewards:
    get:
      description: Retrieve list of all redeemable rewards
      produces:
      - application/json
      responses:
        "200":
          description: Rewards retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RewardResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Get all rewards
      tags:
      - Rewards
    post:
      consumes:
      - application/json
      description: Add a new reward to catalogue (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Reward Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRewardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Reward created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RewardResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new reward
      tags:
      - Rewards
  /api/v1/rewards/{id}:
    delete:
      description: Delete reward by ID (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reward deleted
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Reward not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a reward
      tags:
      - Rewards
    get:
      description: Retrieve details of a specific reward by its ID
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reward retrieved
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RewardResponse'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Reward not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Get reward by ID
      tags:
      - Rewards
    put:
      consumes:
      - application/json
      description: Update reward by ID (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Reward Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRewardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reward updated
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RewardResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Reward not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing reward
      tags:
      - Rewards
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer " followed by your JWT token. Example: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"